# Improve performance of single primitive filter (10%-2.5x)