This improves the performance of filtering primitive arrays, which was using a very basic implementation

```
filter 2^20 f32         time:   [1.6885 ms 1.6976 ms 1.7072 ms]                             
                        change: [-59.388% -59.041% -58.705%] (p = 0.00 < 0.05)

filter null 2^20 f32    time:   [8.5732 ms 8.6265 ms 8.6939 ms]                                 
                        change: [-5.5667% -4.4669% -3.4297%] (p = 0.00 < 0.05)
```

it is interesting that filtering is one of the operations that is the most penalized by Arrow's design of using bitmaps. Filtering a `&[bool]` is much faster than filtering a `Bitmap`.